# api/openapi.yaml
openapi: 3.1.0
info:
  title: SlateDesk API
  description: |
    SlateDesk v2 HTTP API. Milestone 1 surface: local authentication
    (signed HttpOnly session cookie), users, and teams. Milestone 2 adds
    the ticket core: tickets, articles, attachments, tags, dashboard
    counters, and the realtime event stream. Milestone 3 adds the email
    channel's admin surface: mailbox configuration with live fetch/send
    tests and the Google OAuth connect flow, the instance external-URL
    setting, and outbound delivery status on articles with a retry
    operation.

    All errors are RFC 9457 problem details (`application/problem+json`).

    Access model (M2): every ticket-core operation requires an
    authenticated **agent or admin** session — customer access arrives
    with the M4 portal. The one stricter rule: `DELETE /tags/{id}` is
    **admin only**. Agents can additionally **read** users and teams
    (`GET /users`, `/users/{id}`, `/teams`, `/teams/{id}`) — the
    workspace needs them for assignee pickers, team routing, and
    requester lookups; all user/team mutations remain admin only. Each
    operation's description states its required role.

    Access model (M3): everything under `/mailboxes` and `/settings` is
    **admin only** (including the Google OAuth callback, which arrives as
    a top-level browser navigation carrying the admin's session cookie).
    `POST /articles/{id}/retry-send` is agent or admin, like the rest of
...

...
    (`/canned-replies`).

...

...

...
servers:
...
security:
  - cookieAuth: []
  - apiKey: []
...
paths:
...
  /api-keys:
    get:
      operationId: listApiKeys
      summary: List API keys
      description: |
        All API keys, live first then revoked, newest first. Secrets are
        never included — only the display prefix, scopes, and usage
        metadata. **Admin only.**
      tags: [api-keys]
      responses:
        "200":
          description: All API keys.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: createApiKey
      summary: Create an API key
      description: |
        Mints a scoped key. The full plaintext key (`sd_live_…`) is
        returned **once** in this response and never again — store it now.
        Only its SHA-256 digest is persisted. **Admin only.**
      tags: [api-keys]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApiKeyRequest"
      responses:
        "201":
          description: Key created; plaintext returned once.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/Problem"
  /api-keys/{id}:
    parameters:
      - $ref: "#/components/parameters/ApiKeyID"
    delete:
      operationId: revokeApiKey
      summary: Revoke an API key
      description: |
        Revokes the key (soft delete): it stops authenticating
        immediately, but the row is retained for the audit trail.
        Idempotent — revoking an unknown or already-revoked key is a 404.
        **Admin only.**
      tags: [api-keys]
      responses:
        "204":
          description: Key revoked.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
...
  /canned-replies:
    get:
      operationId: listCannedReplies
      summary: List canned replies
      description: |
        All canned replies, sorted by title. **Agent or admin.**
      tags: [canned-replies]
      responses:
        "200":
          description: All canned replies.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CannedReply"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: createCannedReply
      summary: Create a canned reply
      description: |
        Stores a reusable snippet. The `body` keeps its raw template;
        `{{ticket.number}}` and `{{requester.name}}` placeholders are
        substituted at composer-insert time, not stored expanded. **Agent
        or admin.**
      tags: [canned-replies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCannedReplyRequest"
      responses:
        "201":
          description: Canned reply created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CannedReply"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        default:
          $ref: "#/components/responses/Problem"
  /canned-replies/{id}:
    parameters:
      - $ref: "#/components/parameters/CannedReplyID"
    patch:
      operationId: updateCannedReply
      summary: Update a canned reply
      description: |
        Partial update; omitted fields are left unchanged. **Agent or
        admin.**
      tags: [canned-replies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCannedReplyRequest"
      responses:
        "200":
          description: The updated canned reply.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CannedReply"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteCannedReply
      summary: Delete a canned reply
      description: |
        Permanently removes the snippet. **Agent or admin.**
      tags: [canned-replies]
      responses:
        "204":
          description: Canned reply deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        default:
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sd_session
      description: |
        Signed HttpOnly session cookie (SameSite=Lax) carrying the user id
        and token_version. Obtained via POST /auth/login; cleared via
        POST /auth/logout.
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: sd_live_...
      description: |
        Scoped machine API key presented as
        `Authorization: Bearer sd_live_…`. An **alternative** to
        `cookieAuth` on every `/api/v1` operation: because the top-level
        `security` lists the two schemes as separate requirement objects
        (an OR), either scheme alone authenticates a request and the two
        are never both required. The key is a public `sd_live_` prefix
        plus 256 bits of base62 entropy; only its SHA-256 digest is
        stored, and the request acts as the admin or agent who created
        the key, carrying that key's scopes. Scopes are `read` and
        `write` — a `read` key is refused (403) on mutating operations.
        OpenAPI cannot bind scopes to an HTTP-bearer scheme, so the scope
        gate is enforced in the handler and documented rather than
        expressed as security scopes here.
  parameters:
...
    ApiKeyID:
      name: id
      in: path
      required: true
      description: API key id.
      schema:
        type: string
        format: uuid
...
    CannedReplyID:
      name: id
      in: path
      required: true
      description: Canned reply id.
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Malformed request body or parameters.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Missing, invalid, or expired session.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Authenticated but not allowed to perform this action.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: No such resource.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Problem:
      description: Any other error, as RFC 9457 problem details.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details. Media type application/problem+json.
      required: [title, status]
      properties:
        type:
          type: string
          format: uri-reference
          description: Problem type URI; "about:blank" when unspecified.
          default: "about:blank"
        title:
          type: string
          description: Short human-readable summary of the problem type.
        status:
          type: integer
          description: HTTP status code.
          minimum: 100
          maximum: 599
        detail:
          type: string
          description: Occurrence-specific explanation.
        instance:
          type: string
          format: uri-reference
          description: URI reference identifying this occurrence (request path).
...
    ApiKeyScope:
      type: string
      description: |
        API-key permission. `read` allows safe (GET) operations; `write`
        additionally allows mutating operations.
      enum: [read, write]
    ApiKey:
      type: object
      description: |
        Scoped machine API key metadata. The secret is never included —
        only the display prefix and usage columns.
      required:
        - id
        - name
        - key_prefix
        - scopes
        - created_by
        - last_used_at
        - created_at
        - revoked_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        key_prefix:
          type: string
          description: Non-secret display fragment (e.g. "sd_live_ab12cd").
        scopes:
          type: array
          items:
            $ref: "#/components/schemas/ApiKeyScope"
        created_by:
          type: string
          format: uuid
          description: |
            The admin/agent who minted the key; requests act as this user.
        last_used_at:
          type: ["string", "null"]
          format: date-time
          description: |
            Last successful authentication with this key; null if never
            used.
        created_at:
          type: string
          format: date-time
        revoked_at:
          type: ["string", "null"]
          format: date-time
          description: When the key was revoked; null while live.
    CreateApiKeyRequest:
      type: object
      required: [name, scopes]
      properties:
        name:
          type: string
          minLength: 1
        scopes:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: "#/components/schemas/ApiKeyScope"
    ApiKeyCreated:
      description: |
        A newly minted key including its **one-time** plaintext. The `key`
        is shown only in this response and can never be retrieved again.
      allOf:
        - $ref: "#/components/schemas/ApiKey"
        - type: object
          required: [key]
          properties:
            key:
              type: string
              description: |
                Full plaintext key (`sd_live_…`). Returned once; store it
                now. Only its SHA-256 digest is persisted.
...
    CannedReply:
      type: object
      description: |
        A reusable agent reply snippet. `body` stores the raw template;
        `{{ticket.number}}` / `{{requester.name}}` are substituted at
        composer-insert time.
      required: [id, title, body, created_by, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: string
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateCannedReplyRequest:
      type: object
      required: [title, body]
      properties:
        title:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1
    UpdateCannedReplyRequest:
      type: object
      description: Partial update; omitted fields are left unchanged.
      properties:
        title:
          type: string
          minLength: 1
        body:
          type: string
          minLength: 1